separate loaders form routes
